requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
googlemaps==4.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonify==0.5
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
twilio==9.4.2
unittest-xml-reporting==3.2.0
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
